More 1.1 bug fixes
git-svn-id: svn://192.168.1.10/public/Zimodem@14569 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ zimodem/wificlientnode.h
@@ -26,7 +26,7 @@ class WiFiClientNode
     bool wasConnected=false;
     bool serverClient=false;
     uint8 lastPacketBuf[256];
-    uint8 lastPacketLen=0;
+    int lastPacketLen=0;
     WiFiClient *client;
     WiFiClientNode *next = null;
